requirements.txt
fastembed
llama-index
llama-index-embeddings-fastembed
llama-index-embeddings-gemini
llama-index-llms-gemini
llama-index-vector-stores-qdrant
llama-parse
python-dotenv
qdrant_client
streamlit
openpyxl
pandas